Add file that contains stock tickers and names
The file contains the stock tickers,names and industry of all stock in the
Standards & Poor 500 index.

This file will be used to store the tickers of all the stocks that I want to
analyse with my application.
Source: http://data.okfn.org/data/core/s-and-p-500-companies/r/constituents.csv