Add class for storing daily percentage changes
This is a helper class that is used for storing the daily percentage change
for a given date. It contains some helper methods inside it.

This was originally inteded to help with calcualting annualised returns but I
went with a different solution to that problem. This class will definately be
helpful for future calculations.