Fix adding stock to portfolio from stocks view
I added a hidden form that appears when the 'add' symbol is clicked for a given
stock. It allows the user to enter the amount of stock and fill in an optional
buy date. If the buy date is not filled in by the user than the application
assumes that the buy date is $today.