Add checking goals for conflicts
Checking goals for conflicts such as I want 70% of my stock in X and 45% of my
stock in Y would conflict because it is over 100%.
Another conflict that I check for is when I have a goal for having 30% of my
portfolio in Industrials sector, and I try to add a new goal where I want to
have 50% of my portfolo in Industrials. The proper way to do this is by
deleting the old goal, before I can add the new one.